negligible changes, starting up tomorrow morning.
As far as overwriting an entire block, I think the program works based on
writing new blocks that are still valid, meaning they have their own
BlockDetails. I think either way we implement the program the same issue
would exist.

I talked with Garrett after class about abstracting the file implementation
using an array of disks and he made it seem like it was not going to work
but he couldn't convice me why. I'm not against switching to the file
implementation but it would be nice to know a specific reason why its
necessary.