Implented openDisk function. Started work on readBlock.
openDisk seems to be working. There is a global disk entry table that
has information for each disk, including if it is mounted or not. One
thing that is weird is that when using printf instead of fprintf, the
values in the global disk entry table get modified, often leading to
segfaults. I haven't tested this function thoroughly yet.

readBlock is buggy. Right now having trouble where fseek is leading to
segfaults. Will debug some more in the morning.